Do not open remote connection for every single dir
While I was looking at #147 I found a bad design in the current SSH implementation (which has bad effect as well on FTP).

We are opening a connection everytime we go recursively within directories.
We should share the same connection instead: open it once at the beginning of the job and close at the end only.

Closes #159.
(cherry picked from commit 52f2aeb)